+++ android/app/src/main/java/com/on/turip/data/creator/CreatorMapper.kt
@@ -1,7 +1,7 @@
 package com.on.turip.data.creator
 
 import com.on.turip.data.creator.dto.CreatorResponse
-import com.on.turip.domain.videoinfo.contents.creator.Creator
+import com.on.turip.domain.creator.Creator
 
 fun CreatorResponse.toDomain(): Creator =
     Creator(
