+++ android/app/src/main/java/com/on/turip/data/creator/repository/DefaultCreatorRepository.kt
@@ -2,8 +2,8 @@ package com.on.turip.data.creator.repository
 
 import com.on.turip.data.creator.dataSource.CreatorRemoteDataSource
 import com.on.turip.data.creator.toDomain
-import com.on.turip.domain.videoinfo.contents.creator.Creator
-import com.on.turip.domain.videoinfo.contents.creator.repository.CreatorRepository
+import com.on.turip.domain.creator.Creator
+import com.on.turip.domain.creator.repository.CreatorRepository
 
 class DefaultCreatorRepository(
     private val creatorRemoteDataSource: CreatorRemoteDataSource,
