+++ android/app/src/main/java/com/on/turip/data/trip/repository/DefaultTripRepository.kt
@@ -2,8 +2,8 @@ package com.on.turip.data.trip.repository
 
 import com.on.turip.data.trip.dataSource.TripRemoteDataSource
 import com.on.turip.data.trip.toDomain
-import com.on.turip.domain.videoinfo.contents.video.trip.Trip
-import com.on.turip.domain.videoinfo.contents.video.trip.repository.TripRepository
+import com.on.turip.domain.trip.Trip
+import com.on.turip.domain.trip.repository.TripRepository
 
 class DefaultTripRepository(
     private val tripRemoteDataSource: TripRemoteDataSource,
