+++ android/app/src/main/java/com/on/turip/di/RepositoryModule.kt
@@ -3,9 +3,9 @@ package com.on.turip.di
 import com.on.turip.data.content.repository.DefaultContentRepository
 import com.on.turip.data.creator.repository.DefaultCreatorRepository
 import com.on.turip.data.trip.repository.DefaultTripRepository
-import com.on.turip.domain.videoinfo.contents.creator.repository.CreatorRepository
-import com.on.turip.domain.videoinfo.contents.repository.ContentRepository
-import com.on.turip.domain.videoinfo.contents.video.trip.repository.TripRepository
+import com.on.turip.domain.content.repository.ContentRepository
+import com.on.turip.domain.creator.repository.CreatorRepository
+import com.on.turip.domain.trip.repository.TripRepository
 
 object RepositoryModule {
     val contentRepository: ContentRepository by lazy {
