+++ android/app/src/main/java/com/on/turip/domain/content/PagedContentsResult.kt
@@ -1,0 +1,8 @@
+package com.on.turip.domain.content
+
+import com.on.turip.domain.content.video.VideoInformation
+
+data class PagedContentsResult(
+    val videos: List<VideoInformation>,
+    val loadable: Boolean,
+)
