+++ android/app/src/main/java/com/on/turip/domain/content/video/VideoInformation.kt
@@ -1,0 +1,9 @@
+package com.on.turip.domain.content.video
+
+import com.on.turip.domain.content.Content
+import com.on.turip.domain.trip.Trip
+
+data class VideoInformation(
+    val content: Content,
+    val trip: Trip,
+)
