+++ android/app/src/main/java/com/on/turip/domain/creator/repository/CreatorRepository.kt
@@ -1,0 +1,7 @@
+package com.on.turip.domain.creator.repository
+
+import com.on.turip.domain.creator.Creator
+
+interface CreatorRepository {
+    suspend fun loadCreator(creatorId: Long): Result<Creator>
+}
