+++ android/app/src/main/java/com/on/turip/domain/trip/TripCourse.kt
@@ -1,4 +1,4 @@
-package com.on.turip.domain.videoinfo.contents.video.trip
+package com.on.turip.domain.trip
 
 data class TripCourse(
     val tripCourseId: Long,
