+++ android/app/src/main/java/com/on/turip/domain/trip/repository/TripRepository.kt
@@ -1,0 +1,7 @@
+package com.on.turip.domain.trip.repository
+
+import com.on.turip.domain.trip.Trip
+
+interface TripRepository {
+    suspend fun loadTripInfo(contentId: Long): Result<Trip>
+}
