--- android/app/src/main/java/com/on/turip/domain/videoinfo/contents/PagedContentsResult.kt
@@ -1,8 +1,0 @@
-package com.on.turip.domain.videoinfo.contents
-
-import com.on.turip.domain.videoinfo.contents.video.VideoInformation
-
-data class PagedContentsResult(
-    val videos: List<VideoInformation>,
-    val loadable: Boolean,
-)
