--- android/app/src/main/java/com/on/turip/domain/videoinfo/contents/creator/repository/CreatorRepository.kt
@@ -1,7 +1,0 @@
-package com.on.turip.domain.videoinfo.contents.creator.repository
-
-import com.on.turip.domain.videoinfo.contents.creator.Creator
-
-interface CreatorRepository {
-    suspend fun loadCreator(creatorId: Long): Result<Creator>
-}
