+++ android/app/src/main/java/com/on/turip/ui/search/result/SearchResultViewModel.kt
@@ -6,9 +6,9 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import com.on.turip.di.RepositoryModule
-import com.on.turip.domain.videoinfo.contents.PagedContentsResult
-import com.on.turip.domain.videoinfo.contents.repository.ContentRepository
-import com.on.turip.domain.videoinfo.contents.video.VideoInformation
+import com.on.turip.domain.content.PagedContentsResult
+import com.on.turip.domain.content.repository.ContentRepository
+import com.on.turip.domain.content.video.VideoInformation
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
